Move deduplication code to the service
