+++ app/models/user.rb
@@ -224,25 +224,6 @@ class User < ApplicationRecord
     archived_email.present?
   end
 
-  def self.dedup_users!
-    select_users_to_archive.each(&:archive!)
-  end
-
-  def self.select_users_to_archive
-    ids = left_joins(:participant_identities).where(participant_identities: { user_id: nil }).pluck(:id)
-
-    ids_to_exclude = []
-    ids_to_exclude << AdminProfile.where(user_id: ids).pluck(:user_id)
-    ids_to_exclude << AppropriateBodyProfile.where(user_id: ids).pluck(:user_id)
-    ids_to_exclude << DeliveryPartnerProfile.where(user_id: ids).pluck(:user_id)
-    ids_to_exclude << FinanceProfile.where(user_id: ids).pluck(:user_id)
-    ids_to_exclude << InductionCoordinatorProfile.where(user_id: ids).pluck(:user_id)
-    ids_to_exclude << LeadProviderProfile.where(user_id: ids).pluck(:user_id)
-
-    ids -= ids_to_exclude.flatten.uniq
-    User.where(id: ids)
-  end
-
 private
 
   def sync_email_with_identity
